refactor(conn): move to topic mod
