+++ crates/harmion-connect/src/lib.rs
@@ -1,35 +1,8 @@
 // Topic, Connectionなどいろいろ考えれてないことばかり
 
-pub trait Topic {
-    fn value(&self) -> &str;
-}
+mod topic;
 
-impl<T: AsRef<str>> Topic for T {
-    fn value(&self) -> &str {
-        self.as_ref()
-    }
-}
-
-// child, parent
-pub struct TopicTree(pub Vec<Box<dyn Topic>>);
-
-// topic!("/to/file" / harmion::channel::builtin::Write)
-// Write -> harmion/ ...
-// parent: None
-#[macro_export]
-macro_rules! topic {
-    ( $single:expr ) => {{
-        let mut v = Vec::new();
-        v.push(Box::new($single) as Box<dyn Topic>);
-        TopicTree(v)
-    }};
-    // 再帰ケース
-    ( $head:expr , $($rest:expr),+ ) => {{
-        let mut tree = $crate::topic!($($rest),+).0;
-        tree.push(Box::new($head) as Box<dyn Topic>);
-        TopicTree(tree)
-    }};
-}
+use topic::TopicTree;
 
 pub trait Subscriber {
     type E: std::error::Error;
@@ -57,7 +30,7 @@ use serde::{de::DeserializeOwned, Serialize};
 
 pub trait Connection
 where
-    Self: Subscriber + Receiver + Sender,
+    Self: Subscriber + Sender,
 {
     type NodeInfo: Serialize + DeserializeOwned; // SDP
     type NodeID: Serialize + DeserializeOwned; // Node Ident
@@ -67,24 +40,11 @@ where
     fn node_info(&self) -> &(Self::NodeInfo, Self::NodeInfo);
     fn node_list(&self) -> &[(Self::NodeInfo, Self::NodeInfo)];
 
+    fn receiver(&self, topic: &TopicTree) -> impl Receiver;
+
     async fn connect(
         &mut self,
         node: Self::NodeInfo,
         node_id: Option<Self::NodeID>,
     ) -> Result<Self::NodeID, <Self as Connection>::E>;
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::Topic;
-    use crate::TopicTree;
-
-    #[test]
-    fn topic_macro() -> Result<(), Box<dyn std::error::Error>> {
-        let mut tt = super::topic!("harmion", "write").0;
-
-        assert_eq!(tt.len(), 2);
-        assert_eq!(tt.pop().unwrap().value(), "harmion");
-        Ok(())
-    }
 }
